Implement locking for event handlers

+++ SchedulerBot.Client/Program.cs
@@ -29,6 +29,7 @@ using RedLockNet.SERedis.Configuration;
 using System.Net;
 using RedLockNet.SERedis;
 using DSharpPlus.CommandsNext.Exceptions;
+using SchedulerBot.Client.Exceptions;
 
 namespace SchedulerBot.Client
 {
@@ -213,25 +214,69 @@ namespace SchedulerBot.Client
             };
 
             var calendarService = ServiceProvider.GetService<ICalendarService>();
-            await calendarService.CreateCalendarAsync(calendar);
+            var redlockFactory = ServiceProvider.GetService<IDistributedLockFactory>();
+            using (var redlock = await redlockFactory.CreateLockAsync(e.Guild.Id.ToString(), TimeSpan.FromSeconds(15), TimeSpan.FromSeconds(10), TimeSpan.FromSeconds(0.5)))
+            {
+                if (redlock.IsAcquired)
+                {
+                    await calendarService.CreateCalendarAsync(calendar);
+                }
+                else
+                {
+                    throw new RedisLockAcquireException($"Cannot acquire lock for guild {e.Guild.Id}");
+                }
+            }
         }
 
         private async Task OnGuildDelete(GuildDeleteEventArgs e)
         {
             var calendarService = ServiceProvider.GetService<ICalendarService>();
-            await calendarService.DeleteCalendarAsync(e.Guild.Id);
+            var redlockFactory = ServiceProvider.GetService<IDistributedLockFactory>();
+            using (var redlock = await redlockFactory.CreateLockAsync(e.Guild.Id.ToString(), TimeSpan.FromSeconds(15), TimeSpan.FromSeconds(10), TimeSpan.FromSeconds(0.5)))
+            {
+                if (redlock.IsAcquired)
+                {
+                    await calendarService.DeleteCalendarAsync(e.Guild.Id);
+                }
+                else
+                {
+                    throw new RedisLockAcquireException($"Cannot acquire lock for guild {e.Guild.Id}");
+                }
+            }
         }
 
         private async Task OnGuildMemberRemove(GuildMemberRemoveEventArgs e)
         {
             var permissionService = ServiceProvider.GetService<IPermissionService>();
-            await permissionService.RemoveUserPermissionsAsync(e.Guild.Id, e.Member.Id);
+            var redlockFactory = ServiceProvider.GetService<IDistributedLockFactory>();
+            using (var redlock = await redlockFactory.CreateLockAsync(e.Guild.Id.ToString(), TimeSpan.FromSeconds(15), TimeSpan.FromSeconds(10), TimeSpan.FromSeconds(0.5)))
+            {
+                if (redlock.IsAcquired)
+                {
+                    await permissionService.RemoveUserPermissionsAsync(e.Guild.Id, e.Member.Id);
+                }
+                else
+                {
+                    throw new RedisLockAcquireException($"Cannot acquire lock for guild {e.Guild.Id}");
+                }
+            }
         }
 
         private async Task OnGuildRoleDelete(GuildRoleDeleteEventArgs e)
         {
             var permissionService = ServiceProvider.GetService<IPermissionService>();
-            await permissionService.RemoveRolePermissionsAsync(e.Guild.Id, e.Role.Id);
+            var redlockFactory = ServiceProvider.GetService<IDistributedLockFactory>();
+            using (var redlock = await redlockFactory.CreateLockAsync(e.Guild.Id.ToString(), TimeSpan.FromSeconds(15), TimeSpan.FromSeconds(10), TimeSpan.FromSeconds(0.5)))
+            {
+                if (redlock.IsAcquired)
+                {
+                    await permissionService.RemoveRolePermissionsAsync(e.Guild.Id, e.Role.Id);
+                }
+                else
+                {
+                    throw new RedisLockAcquireException($"Cannot acquire lock for guild {e.Guild.Id}");
+                }
+            }
         }
 
         private async Task OnClientReady(ReadyEventArgs e)
